- 추가(ConnectionManager): 탭아이디-커넥션, 아티클-커넥션 2가지 커넥션을 관리하는 클래스 추가     - 커넥션 관련 작업은 모두 이 클래스에서 수행하도록 변경     - 기존의 콜백을 등록하는 코드를 이 클래스에 내재화하여 서비스 코드 간소화
